% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cams_solar_radiation.R
\name{cams_solar_radiation}
\alias{cams_solar_radiation}
\title{R6 Class for CAMS solar radiation}
\description{
description
}
\details{
details
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{locations}}{Get all streams}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{cams_solar_radiation$new()}}
\item \href{#method-add}{\code{cams_solar_radiation$add()}}
\item \href{#method-download}{\code{cams_solar_radiation$download()}}
\item \href{#method-process_l0_to_l1}{\code{cams_solar_radiation$process_l0_to_l1()}}
\item \href{#method-read}{\code{cams_solar_radiation$read()}}
\item \href{#method-get}{\code{cams_solar_radiation$get()}}
\item \href{#method-update}{\code{cams_solar_radiation$update()}}
\item \href{#method-clone}{\code{cams_solar_radiation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cams_solar_radiation$new(
  dir_db = "databases/CAMS",
  dir_l0 = "l0/solar",
  dir_l1 = "l1/solar",
  init_db = FALSE,
  init_locations = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cams_solar_radiation$add(place, lat, lon, alt)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download"></a>}}
\if{latex}{\out{\hypertarget{method-download}{}}}
\subsection{Method \code{download()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cams_solar_radiation$download(place, from = "2005-01-01", to = Sys.Date())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-process_l0_to_l1"></a>}}
\if{latex}{\out{\hypertarget{method-process_l0_to_l1}{}}}
\subsection{Method \code{process_l0_to_l1()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cams_solar_radiation$process_l0_to_l1(place, save = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\if{latex}{\out{\hypertarget{method-read}{}}}
\subsection{Method \code{read()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cams_solar_radiation$read(
  place,
  year_max = lubridate::year(Sys.Date()),
  from = NULL,
  to = NULL,
  scale_GHI = 1000
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cams_solar_radiation$get(
  place,
  lat,
  lon,
  alt,
  from = "2005-01-01",
  to = Sys.Date()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cams_solar_radiation$update(place)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cams_solar_radiation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
