% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cams_solar_radiation_internals.R
\name{cams_solar_radiation_process_l0_to_l1}
\alias{cams_solar_radiation_process_l0_to_l1}
\title{Process raw data into a clean dataset}
\usage{
cams_solar_radiation_process_l0_to_l1(place, data, meta)
}
\description{
Process raw data into a clean dataset
}
\keyword{internal}
