% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cams_solar_radiation_internals.R
\name{cams_solar_radiation_ts_download}
\alias{cams_solar_radiation_ts_download}
\title{Get a time series of solar radiation from CAMS Api directly as `tibble` without the necessity of store `.csv` files.
Call the python function `cams_solar_radiation_ts`.}
\usage{
cams_solar_radiation_ts_download(
  place,
  lat,
  lon,
  alt,
  from = "2005-01-01",
  to = Sys.Date()
)
}
\description{
Get a time series of solar radiation from CAMS Api directly as `tibble` without the necessity of store `.csv` files.
Call the python function `cams_solar_radiation_ts`.
}
\keyword{internal}
